stable-compat: Device in device.h (==/is_cpu), flatten, cublas_v2.h shim
More torch-2.10 stable surface newer libtorch-stable kernels use: move Device into torch/csrc/stable/device.h with operator==/!= and is_cpu(); add torch::stable::flatten (aten::flatten.using_ints via the dispatcher); add a cublas_v2.h -> mublas.h shim.
